Add 18F/sails-postgres to package.json. Reorganize configs.
