+++ .travis.yml
@@ -3,30 +3,3 @@ node_js:
   - "0.10"
 
 before_install:
-  - mkdir node_modules
-  - git clone git://github.com/Innovation-Toolkit/sails-postgresql.git
-  - cd sails-postgresql
-  - git checkout bytea
-  - npm install
-  - cd ../node_modules
-  - ln -s ../sails-postgresql sails-postgresql
-  - cd ..
-  # - git clone git://github.com/Innovation-Toolkit/anchor.git
-  # - cd anchor
-  # - npm install
-  # - cd ..
-  # - git clone git://github.com/balderdashy/sails.git
-  # - cd sails
-  # - git checkout 88023b5e4557b42258e6f651b99070d19d55502c
-  # - mkdir node_modules
-  # - cd node_modules
-  # - ln -s ../../anchor anchor
-  # - cd ..
-  # - ls -l node_modules
-  # - npm install
-  # - cd ../node_modules
-  # - ln -s ../anchor anchor
-  # - ln -s ../sails sails
-  # - cd ..
-  # - ls -l node_modules
-  # - pwd
