+++ config/bootstrap.js
@@ -20,6 +20,6 @@ module.exports.bootstrap = function (cb) {
            sails.services = modules;
            cb();
          });
-  // It's very important to trigger this callback method when you are finished 
+  // It's very important to trigger this callback method when you are finished
   // with the bootstrap!  (otherwise your server will never lift, since it's waiting on the bootstrap)
 };
