+++ config/i18next.js
