+++ config/settings/analytics.js
