+++ config/settings/auth.js
