+++ config/settings/notifications.js
